Updated SMToast functionality
Added two queues to store SMToasts. Added the ability
to present multiple toasts at once, without overlaying
already presented toasts. If more are trying to be presented
than can fit on the screen, they will be added to the "onHold"
queue. On hold toasts will be presented one by one as
toasts disappear from the screen.

Another improvement for SMToast is the ability to drag
them around the screen. This allows users to move
SMToast's out of the way if covering an important piece
of information.